# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SolveX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library/{productId}/check:
    get:
      operationId: checkPurchased
      tags: [library]
      summary: Check if user has purchased a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purchase status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseCheck"

  # ── Problems ──────────────────────────────────────────────────────────
...
  /escrow:
    post:
      operationId: createEscrow
      tags: [escrow]
      summary: Create escrow for a problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscrowInput"
      responses:
        "201":
          description: Escrow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escrow"
  /escrow/payment-intent:
    post:
      operationId: createPaymentIntent
      tags: [escrow]
      summary: Create Stripe payment intent for escrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentIntentInput"
      responses:
        "200":
          description: Payment intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentResult"
  /escrow/problem/{problemId}:
    get:
      operationId: getEscrowByProblem
      tags: [escrow]
      summary: Get escrow for a problem
      parameters:
        - name: problemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Escrow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escrow"

  # ── Crawler ───────────────────────────────────────────────────────────
  /crawler/run:
    post:
      operationId: runCrawler
      tags: [crawler]
      summary: Run AI crawler (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrawlerInput"
      responses:
        "200":
          description: Crawler results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerResult"
  /crawler/crawled:
    get:
      operationId: listCrawledProblems
      tags: [crawler]
      summary: List crawled problems (admin only)
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Crawled problems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrawledProblem"
  /crawler/import:
    post:
      operationId: importCrawledProblem
      tags: [crawler]
      summary: Import a crawled problem (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportInput"
      responses:
        "200":
          description: Imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  # ── Notifications ─────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
...
  /offers:
    post:
      operationId: createOffer
      tags: [offers]
      summary: Make an offer on a problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/mine:
    get:
      operationId: getMyOffers
      tags: [offers]
      summary: Get user's offers
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [sent, received]
            default: received
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
  /offers/problem/{problemId}:
    get:
      operationId: getOffersByProblem
      tags: [offers]
      summary: Get offers for a problem
      parameters:
        - name: problemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
  /offers/{id}/counter:
    post:
      operationId: counterOffer
      tags: [offers]
      summary: Counter an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterOfferInput"
      responses:
        "200":
          description: Counter offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}/accept:
    post:
      operationId: acceptOffer
      tags: [offers]
      summary: Accept an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /offers/{id}/reject:
    post:
      operationId: rejectOffer
      tags: [offers]
      summary: Reject an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  # ── Owner Dashboard ───────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    PurchaseCheck:
      type: object
      required: [purchased]
      properties:
        purchased:
          type: boolean
...
    Escrow:
      type: object
      required: [id, problemId, clientId, amount, status]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        clientId:
          type: integer
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [pending, held, released, refunded, failed]
        stripePaymentIntentId:
          type: ["string", "null"]
        releasedAt:
          type: ["string", "null"]
        refundedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    EscrowInput:
      type: object
      required: [problemId, amount]
      properties:
        problemId:
          type: integer
        amount:
          type: number
    PaymentIntentInput:
      type: object
      required: [problemId, amount]
      properties:
        problemId:
          type: integer
        amount:
          type: number
    PaymentIntentResult:
      type: object
      required: [clientSecret, paymentIntentId]
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
    CrawlerInput:
      type: object
      properties:
        platforms:
          type: array
          items:
            type: string
        category:
          type: string
        limit:
          type: integer
          default: 10
    CrawlerResult:
      type: object
      required: [found]
      properties:
        found:
          type: integer
        problems:
          type: array
          items:
            $ref: "#/components/schemas/CrawledProblem"
    CrawledProblem:
      type: object
      required: [id, title, description, platform]
      properties:
        id:
          type: integer
        externalId:
          type: string
        platform:
          type: string
        title:
          type: string
        description:
          type: string
        sourceUrl:
          type: string
        suggestedCategory:
          type: string
        suggestedPayment:
          type: ["string", "null"]
        isImported:
          type: boolean
        aiSummary:
          type: ["string", "null"]
        upvotes:
          type: ["integer", "null"]
        crawledAt:
          type: string
    ImportInput:
      type: object
      required: [crawledId, paymentOffer]
      properties:
        crawledId:
          type: integer
        paymentOffer:
          type: number
...
    Offer:
      type: object
      required: [id, problemId, fromUserId, toUserId, amount, status]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        amount:
          type: string
        currency:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, countered]
        counterOfferId:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    OfferInput:
      type: object
      required: [problemId, amount]
      properties:
        problemId:
          type: integer
        amount:
          type: number
        message:
          type: string
        expiresAt:
          type: string
    CounterOfferInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        message:
          type: string
...